Remove parameter callbacks system; add "refresh" and "monitor" flags
Previously, parameters could be assigned callbacks (such as triggering an Environment snapshot). This was rarely used and sometimes led to unpredictable behavior, so it has been removed.

The "refresh" flag is passed to a Parameter's snapshot() method to determine whether a new measurement should be made, or if it should just return the last value.

The "monitor" flag is a permanent attribute serving the same functionality as "refresh". Whether a new measurement is made is determined by a logical AND between the "refresh" and "monitor" flags.